El método getData de magnetometer devuelve un vector de float,no de int.

+++ magnetometer.cpp
@@ -33,7 +33,7 @@ Magnetometer::Magnetometer() :
 
 }
 
-bool Magnetometer::getData(Vector<int> &vector){
+bool Magnetometer::getData(Vector<float> &vector){
 	if(dataReady()){
 		//The pointer to the registers of device always point to REG_XH
 		//at this point, so do not need move it. We read the 6 bytes directly
